:tractor: fusing engines to template.py
