+++ moban/plugins/template.py
@@ -1,10 +1,36 @@
 import os
 
-from moban import utils, reporter, exceptions
+from lml.plugin import PluginManager
+
+from moban import utils, reporter, constants, exceptions
 from moban.hashstore import HASH_STORE
 from moban.plugins.context import Context
 from moban.plugins.library import LIBRARIES
 from moban.plugins.strategy import Strategy
+
+
+class TemplateFactory(PluginManager):
+    def __init__(self):
+        super(TemplateFactory, self).__init__(
+            constants.TEMPLATE_ENGINE_EXTENSION
+        )
+        self.extensions = {}
+
+    def register_extensions(self, extensions):
+        self.extensions.update(extensions)
+
+    def get_engine(self, template_type, template_dirs, context_dirs):
+        engine_cls = self.load_me_now(template_type)
+        engine_extensions = self.extensions.get(template_type)
+        return TemplateEngine(
+            template_dirs, context_dirs, engine_cls, engine_extensions
+        )
+
+    def all_types(self):
+        return list(self.registry.keys())
+
+    def raise_exception(self, key):
+        raise exceptions.NoThirdPartyEngine(key)
 
 
 class TemplateEngine(object):
