+++ tests/test_jinja2_extensions.py
@@ -3,8 +3,8 @@ import os
 from nose.tools import eq_
 
 from moban.jinja2.engine import Engine
+from moban.plugins.template import TemplateEngine
 from moban.jinja2.extensions import jinja_global
-from moban.plugins.template_engine import TemplateEngine
 
 
 def test_globals():
